Revert "Landing Page Is Mostly Done"
