+++ css/account_creation.css
@@ -33,7 +33,7 @@ body {
   background: rgba(255, 255, 255, 0.65);
   color: black;
 }
- 
+
 .doubleinput {
   width:85%;
 }
