+++ css/style.css
@@ -45,8 +45,7 @@ body{
 
 h1 {
     color: #48456b;
-    padding:5px;
-    
+    margin-bottom: 10px;
 }
 
 h2 {
@@ -83,11 +82,8 @@ p,a{
     padding:15px 80px;
     color: white;
     font-family: Baloo Bhai, sans-serif;
-
-    margin:20px;
-    bottom:0px;
-    right:0;
-    position: absolute;
+	margin-left: 20px;
+	margin-right: 20px;
     opacity: 0.9;
     font-size: 20px;
 	font-weight: bold;
@@ -106,6 +102,12 @@ p,a{
 #backPage .purple_button{
 	top: 50px;
 }
+
+#adpage {
+    background-color: #e4e2f7
+}
+
+
