--- public/testUsers.js
@@ -1,12 +1,0 @@
-const user = {
-	user:'user',
-	password:'u'
-}
-const admin = {
-	user:'admin',
-	password:'a'
-}
-const super = {
-	user:'super',
-	password:'s'
-}
