+++ views/admin_edit.hbs
@@ -7,7 +7,7 @@
 </head>
 
 <body>
-	{{> navbar_logout}}
+	{{> navbar}}
 	<div id='main'>
 		<div id ='options'>
             <div id="parameters">
