+++ views/admin_list_page_edit.hbs
@@ -8,7 +8,7 @@
 </head>
 
 <body>
-    {{> navbar_logout}}
+    {{> navbar}}
 
     <div id='main'>
         <div id='options'>
