+++ views/dashboard.hbs
@@ -10,7 +10,7 @@
 
 <body>
 
-    {{> navbar_logout}}
+    {{> navbar}}
     {{> navbar2}}
