+++ views/partials/loginNavbar.hbs
@@ -7,7 +7,7 @@
     </div>
 
     <div id='right'>
-    	<a id='login' href=./login>
+    	<a id='login' href=./provider_login>
     		Login
     	</a>
     </div>
