--- views/partials/navbar_logout.hbs
@@ -1,14 +1,0 @@
-<div id='navbar'>
-    <div id='left'>
-    </div>
-
-    <div id='center'>
-    	edify Providers
-    </div>
-
-    <div id='right'>
-    	<a id='login'>
-    		Logout
-    	</a>
-    </div>
-</div>
