+++ views/provider_edit.hbs
@@ -10,7 +10,7 @@
 
 <body>
 
-	{{> navbar_logout}}
+	{{> navbar}}
 
 	<div id='main'>
 		<div id='info'>
