+++ views/provider_list.hbs
@@ -6,7 +6,7 @@
 </head>
 
 <body>
-	{{> navbar_logout}}
+	{{> navbar}}
 
 	<div id='main'>
 		<div id ='options'>
