+++ views/settings.hbs
@@ -11,7 +11,7 @@
 </head>
 
 <body id="body">
-    {{> navbar_logout}}
+     {{> navbar}}
     {{> navbar2}}
 
     <div id="maincontentset" >
