+++ views/status.hbs
@@ -11,9 +11,7 @@
 
 <body id="body">
      
-
-    {{> navbar_logout}}
-
+    {{> navbar}}
     {{> navbar2}}
